+++ tests/testthat/test-mcmc-derive.R
@@ -1,8 +1,33 @@
 context("mcmc-derive")
 
-test_that("mcmc_derive", {
+test_that("mcmc_derive.nlist", {
+  nlist <- nlist::nlist(x = 1:2, y = matrix(1:4, 2))
   
-  mcmcr <- subset(mcmcr::mcmcr_example, 1:2, 1:10)
+  expr <- "
+    gamma <- x + 2
+    z <- y[1,2]
+  "
+
+  expect_identical(mcmc_derive(nlist, expr, silent = TRUE),
+                   structure(list(gamma = c(3, 4), z = 3L), class = "nlist"))
+})
+
+test_that("mcmc_derive.nlists", {
+  nlist <- nlist::nlists(nlist::nlist(x = 1:2, y = matrix(1:4, 2)),
+                         nlist::nlist(x = 3:4, y = matrix(4:1, 2)))
+  
+  expr <- "
+    gamma <- x + 2
+    z <- y[1,2]
+  "
+
+  expect_identical(mcmc_derive(nlist, expr, silent = TRUE),
+                   structure(list(structure(list(gamma = c(3, 4), z = 3L), class = "nlist"), 
+    structure(list(gamma = c(5, 6), z = 2L), class = "nlist")), class = "nlists"))
+})
+
+test_that("mcmc_derive.mcmc", {
+  mcmc <- coda::as.mcmc(subset(mcmcr::mcmcr_example, 1L, 1:2))
   
   expr <- "
     gamma <- alpha + beta
@@ -15,53 +40,88 @@ test_that("mcmc_derive", {
   
   values <- list(x = 2:10)
   
-  derived <- mcmc_derive(mcmcr, expr, values = values, monitor = "^g|^a", silent = TRUE)
-  
-  expect_identical(pars(derived), c("alpha2", "alpha3", "gamma"))
-  expect_identical(nchains(derived), 2L)
-  expect_identical(niters(derived), 10L)
-  expect_identical(nterms(derived), 15L)
-  
-  expect_equal(estimates(derived),
-               list(alpha2 = c(6.0195, 8.0195), 
-                    alpha3 = c(6.0195, 9.02925, 12.039, 15.04875, 18.0585, 
-                               21.06825, 24.078, 27.08775, 30.0975), 
-                    gamma = structure(c(3.941852, 5.941852, 4.941852, 6.941852),
-                                      .Dim = c(2L, 2L))))
-  
-  expect_error(mcmc_derive(mcmcr, expr, values = values, monitor = "something",
-                           silent = TRUE), 
-               "monitor 'something' must match at least one of the following variables in expr: 'gamma', 'alpha2', 'znot', 'i' or 'alpha3'")
+  expect_equal(mcmc_derive(mcmc, expr, values = values, monitor = "^g|^a", silent = TRUE),
+                   structure(c(14.34626, 5.90506, 16.34626, 7.90506, 14.34626, 5.90506, 
+21.51939, 8.85759, 28.69252, 11.81012, 35.86565, 14.76265, 43.03878, 
+17.71518, 50.21191, 20.66771, 57.38504, 23.62024, 64.55817, 26.57277, 
+71.7313, 29.5253, 5.60693, 4.4559, 7.60693, 6.4559, 6.60693, 
+5.4559, 8.60693, 7.4559), .Dim = c(2L, 15L), .Dimnames = list(
+    NULL, c("alpha2[1]", "alpha2[2]", "alpha3[1]", "alpha3[2]", 
+    "alpha3[3]", "alpha3[4]", "alpha3[5]", "alpha3[6]", "alpha3[7]", 
+    "alpha3[8]", "alpha3[9]", "gamma[1,1]", "gamma[2,1]", "gamma[1,2]", 
+    "gamma[2,2]")), mcpar = c(1, 2, 1), class = "mcmc"))
 })
 
 test_that("mcmc_derive.mcmc.list", {
-  mcmc_list <- coda::as.mcmc.list(subset(mcmcr::mcmcr_example, 1:2, 1:10))
+  mcmc.list <- coda::as.mcmc.list(subset(mcmcr::mcmcr_example, 1L, 1:2))
   
-  expect_warning(mcmc_derive(mcmc_list, "gamma <- alpha + beta"), 
-                 "the following parameter was not in expr and so was dropped from object: 'sigma'")
-  derived <- mcmc_derive(mcmc_list, "gamma <- alpha + beta", silent = TRUE)
-  expect_identical(pars(derived), "gamma")
-  expect_identical(nchains(derived), 2L)
-  expect_identical(niters(derived), 10L)
-  expect_identical(nterms(derived), 4L)
+  expr <- "
+    gamma <- alpha + beta
+  alpha2 <- alpha * 2
+  znot <- alpha * 2
+  for(i in seq_along(x)) {
+    alpha3[i] <- alpha[1] * x[i]
+  }
+  "
+  
+  values <- list(x = 2:10)
+  
+  expect_equal(mcmc_derive(mcmc.list, expr, values = values, monitor = "^g|^a", silent = TRUE),
+                   structure(list(structure(c(14.34626, 5.90506, 16.34626, 7.90506, 
+14.34626, 5.90506, 21.51939, 8.85759, 28.69252, 11.81012, 35.86565, 
+14.76265, 43.03878, 17.71518, 50.21191, 20.66771, 57.38504, 23.62024, 
+64.55817, 26.57277, 71.7313, 29.5253, 5.60693, 4.4559, 7.60693, 
+6.4559, 6.60693, 5.4559, 8.60693, 7.4559), .Dim = c(2L, 15L), .Dimnames = list(
+    NULL, c("alpha2[1]", "alpha2[2]", "alpha3[1]", "alpha3[2]", 
+    "alpha3[3]", "alpha3[4]", "alpha3[5]", "alpha3[6]", "alpha3[7]", 
+    "alpha3[8]", "alpha3[9]", "gamma[1,1]", "gamma[2,1]", "gamma[1,2]", 
+    "gamma[2,2]")), mcpar = c(1, 2, 1), class = "mcmc")), class = "mcmc.list"))
 })
 
 test_that("mcmc_derive.mcmcr", {
-  mcmcr_example <- subset(mcmcr::mcmcr_example, 1:2, 1:10)
-  mcmcrs <- mcmcr::as.mcmcrs(list(model1 = mcmcr_example, model2 = mcmcr_example))
   
-  expect_warning(mcmc_derive(mcmcrs, "gamma <- alpha + beta"), 
-                 "the following parameter was not in expr and so was dropped from object: 'sigma'")
-  derived <- mcmc_derive(mcmcrs, "gamma <- alpha + beta", silent = TRUE)
-  expect_identical(pars(derived), "gamma")
-  expect_identical(nchains(derived), 2L)
-  expect_identical(niters(derived), 10L)
-  expect_identical(nterms(derived), 4L)
+  mcmcr <- subset(mcmcr::mcmcr_example, 1:2, 1:2)
+  
+  expr <- "
+    gamma <- alpha + beta
+  alpha2 <- alpha * 2
+  znot <- alpha * 2
+  for(i in seq_along(x)) {
+    alpha3[i] <- alpha[1] * x[i]
+  }
+  "
+  
+  values <- list(x = 2:10)
+  
+  expect_equal(mcmc_derive(mcmcr, expr, values = values, monitor = "^g|^a", silent = TRUE),
+                   structure(list(alpha2 = structure(c(14.34626, 4.133, 5.90506, 
+5.10146, 16.34626, 6.133, 7.90506, 7.10146), .Dim = c(2L, 2L, 
+2L), class = "mcmcarray"), alpha3 = structure(c(14.34626, 4.133, 
+5.90506, 5.10146, 21.51939, 6.1995, 8.85759, 7.65219, 28.69252, 
+8.266, 11.81012, 10.20292, 35.86565, 10.3325, 14.76265, 12.75365, 
+43.03878, 12.399, 17.71518, 15.30438, 50.21191, 14.4655, 20.66771, 
+17.85511, 57.38504, 16.532, 23.62024, 20.40584, 64.55817, 18.5985, 
+26.57277, 22.95657, 71.7313, 20.665, 29.5253, 25.5073), .Dim = c(2L, 
+2L, 9L), class = "mcmcarray"), gamma = structure(c(5.60693, 3.01633, 
+4.4559, 3.084598, 7.60693, 5.01633, 6.4559, 5.084598, 6.60693, 
+4.01633, 5.4559, 4.084598, 8.60693, 6.01633, 7.4559, 6.084598
+), .Dim = c(2L, 2L, 2L, 2L), class = "mcmcarray")), class = "mcmcr"))
+})
+
+test_that("mcmc_derive.mcmcrs", {
+  mcmcr <- subset(mcmcr::mcmcr_example, 1L, 1L)
+  mcmcrs <- mcmcr::as.mcmcrs(list(mcmcr, mcmcr))
+  
+  expect_equal(mcmc_derive(mcmcrs, "gamma <- alpha + beta", silent = TRUE),
+                   structure(list(mcmcr1 = structure(list(gamma = structure(c(5.60693, 
+7.60693, 6.60693, 8.60693), .Dim = c(1L, 1L, 2L, 2L), class = "mcmcarray")), class = "mcmcr"), 
+    mcmcr2 = structure(list(gamma = structure(c(5.60693, 7.60693, 
+    6.60693, 8.60693), .Dim = c(1L, 1L, 2L, 2L), class = "mcmcarray")), class = "mcmcr")), class = "mcmcrs"))
 })
 
 test_that("mcmc_derive in parallel", {
   
-  mcmcr <- subset(mcmcr::mcmcr_example, 1:2, 1:10)
+  mcmcr <- subset(mcmcr::mcmcr_example, 1:2, 1:2)
   
   expr <- "
     gamma <- alpha + beta
@@ -76,30 +136,24 @@ test_that("mcmc_derive in parallel", {
   
   doParallel::registerDoParallel(2)
   
-  derived <- mcmc_derive(mcmcr, expr, values = values, monitor = "^g|^a", 
-                         parallel = TRUE,silent = TRUE)
-  
-  expect_identical(pars(derived), c("alpha2", "alpha3", "gamma"))
-  expect_identical(nchains(derived), 2L)
-  expect_identical(niters(derived), 10L)
-  expect_identical(nterms(derived), 15L)
-  
-  
-  expect_equal(estimates(derived),
-               list(alpha2 = c(6.0195, 8.0195), 
-                    alpha3 = c(6.0195, 9.02925, 12.039, 15.04875, 18.0585, 
-                               21.06825, 24.078, 27.08775, 30.0975), 
-                    gamma = structure(c(3.941852, 5.941852, 4.941852, 6.941852),
-                                      .Dim = c(2L, 2L))))
-  
-  expect_error(mcmc_derive(mcmcr, expr, values = values, monitor = "something",
-                           silent = TRUE), 
-               "monitor 'something' must match at least one of the following variables in expr: 'gamma', 'alpha2', 'znot', 'i' or 'alpha3'")
+  expect_equal(mcmc_derive(mcmcr, expr, values = values, monitor = "^g|^a", 
+                         parallel = TRUE, silent = TRUE),
+  structure(list(alpha2 = structure(c(14.34626, 4.133, 5.90506, 
+5.10146, 16.34626, 6.133, 7.90506, 7.10146), .Dim = c(2L, 2L, 
+2L), class = "mcmcarray"), alpha3 = structure(c(14.34626, 4.133, 
+5.90506, 5.10146, 21.51939, 6.1995, 8.85759, 7.65219, 28.69252, 
+8.266, 11.81012, 10.20292, 35.86565, 10.3325, 14.76265, 12.75365, 
+43.03878, 12.399, 17.71518, 15.30438, 50.21191, 14.4655, 20.66771, 
+17.85511, 57.38504, 16.532, 23.62024, 20.40584, 64.55817, 18.5985, 
+26.57277, 22.95657, 71.7313, 20.665, 29.5253, 25.5073), .Dim = c(2L, 
+2L, 9L), class = "mcmcarray"), gamma = structure(c(5.60693, 3.01633, 
+4.4559, 3.084598, 7.60693, 5.01633, 6.4559, 5.084598, 6.60693, 
+4.01633, 5.4559, 4.084598, 8.60693, 6.01633, 7.4559, 6.084598
+), .Dim = c(2L, 2L, 2L, 2L), class = "mcmcarray")), class = "mcmcr"))
 })
 
-
 test_that("mcmc_derive matrix in values", {
-  mcmcr <- subset( mcmcr::mcmcr_example, 1:2, 1:10, pars = "beta")
+  mcmcr <- subset( mcmcr::mcmcr_example, 1:2, 1:2, pars = "beta")
   
   expr <- "
     x <- Z
@@ -114,14 +168,14 @@ test_that("mcmc_derive matrix in values", {
   
   values <- list(Z = Z)
   
-  derived <- mcmc_derive(mcmcr, expr, values = values, monitor = "x")
-  
-  names(derived) <- "beta"
-  expect_identical(derived, mcmcr)
+  expect_equal(mcmc_derive(mcmcr, expr, values = values, monitor = "x"),
+                   structure(list(x = structure(c(-1.5662, 0.94983, 1.50337, 0.533868, 
+-0.5662, 1.94983, 2.50337, 1.533868, -0.5662, 1.94983, 2.50337, 
+1.533868, 0.4338, 2.94983, 3.50337, 2.533868), .Dim = c(2L, 2L, 
+2L, 2L), class = "mcmcarray")), class = "mcmcr"))
 })
 
 test_that("mcmc_derive warnings and errors", {
-  
   mcmcr <- subset(mcmcr::mcmcr_example, 1:2, 1:10)
   
   expect_error(mcmc_derive(mcmcr, expr = "garbage r code"),
@@ -158,4 +212,8 @@ test_that("mcmc_derive warnings and errors", {
   
   expect_error(mcmc_derive(mcmcr, expr = "alpha2 <- beta * sigma * alpha * alpha3", monitor = "2$"),
                "the following derived parameter includes missing values: 'alpha2'")
+
+  expect_error(mcmc_derive(mcmcr, expr = "gamma <- alpha", monitor = "something",
+                           silent = TRUE),
+               "monitor 'something' must match at least one of the following variable in expr: 'gamma'")
 })