+++ NEWS.md
@@ -1,3 +1,4 @@
+- Added mcmc_derive() for nlist, nlists and mcmc objects
 - Internal
   - Replace parameters() with pars()
   - Replace subset_mcmcr_parameters() with subset()